refactor: staged_ledger_diff.rs with coinbase.rs
