+++ base/tests/staged_ledger_diff.rs
@@ -7,11 +7,11 @@ wasm_bindgen_test::wasm_bindgen_test_configure!(run_in_browser);
 #[cfg(test)]
 mod tests {
     use mina_crypto::base58::Base58Encodable;
+    use mina_rs_base::coinbase::*;
     use mina_rs_base::snapp_command::SnappCommand;
-    use mina_rs_base::staged_ledger_diff::{
-        CoinBase, CoinBaseBalanceData, CoinBaseFeeTransfer, FeeTransferBalanceData,
+    use mina_rs_base::transaction_status::{
+        CoinBaseBalanceData, FeeTransferBalanceData, TransactionStatus,
     };
-    use mina_rs_base::transaction_status::TransactionStatus;
     use mina_rs_base::types::*;
     use time::macros::*;
     use wasm_bindgen_test::*;
